- Added URL support to "cat"; like cat http://192.168.0.1/ - Removed some warnings - Fixed permission issue in http protocol implementation
git-svn-id: svn+ssh://svn.code.sf.net/p/jnode/svn/trunk@302 407389f7-7c16-0410-9eee-ad85acb55d7f

+++ net/src/net/org/jnode/net/ipv4/dhcp/DHCPClient.java
@@ -56,7 +56,7 @@ public class DHCPClient extends BOOTPClient {
 					if(value.length == 1)
 						log.debug("Option "+n+" : " + (int)(value[0]));
 					else if(value.length == 2)
-						log.debug("Option "+n+" : " + (int)((value[0] << 8) | value[1]));
+						log.debug("Option "+n+" : " + ((value[0] << 8) | value[1]));
 					else if(value.length == 4)
 						log.debug("Option "+n+" : " + InetAddress.getByAddress(value).toString());
 					else
